requirements.txt
Django==3.2.7
django-cors-headers==3.8.0
django-rest-framework==0.1.0
django-uwsgi==0.2.2
djangorestframework==3.12.4
lxml==4.5.0
psycopg2-binary==2.8.6
requests==2.18.4
requests-oauthlib==1.3.0
urllib3==1.22
uWSGI==2.0.17
uwsgi-tasks==0.7.2
